refactor(iif): resolve issues introduced by last rebase
